Fixes the Screen Format screen to have friendly names
